Make fec's silly 'magic-header' stuff work on 64-bit platforms.
I think.

+++ src/fec.c
@@ -1,6 +1,6 @@
 /* Portions Copyright (c) 2003 Nick Mathewson.  See LICENCE for licensing
  * information. */
-/* $Id: fec.c,v 1.11 2003/11/28 04:14:05 nickm Exp $ */
+/* $Id: fec.c,v 1.12 2003/12/12 23:53:31 nickm Exp $ */
 
 #include <Python.h>
 #include "_minionlib.h"
@@ -573,7 +573,7 @@ static void
 fec_free(struct fec_parms *p)
 {
     if (p==NULL ||
-       p->magic != ( ( (FEC_MAGIC ^ p->k) ^ p->n) ^ (int)(p->enc_matrix)) ) {
+       p->magic != (((FEC_MAGIC ^ p->k) ^ p->n) ^ (unsigned long)(p->enc_matrix))) {
 	fprintf(stderr, "bad parameters to fec_free\n");
 	return ;
     }
@@ -605,7 +605,7 @@ fec_new(int k, int n)
     retval->k = k ;
     retval->n = n ;
     retval->enc_matrix = NEW_GF_MATRIX(n, k);
-    retval->magic = ( ( FEC_MAGIC ^ k) ^ n) ^ (int)(retval->enc_matrix) ;
+    retval->magic = ((FEC_MAGIC ^ k) ^ n) ^ (unsigned long)(retval->enc_matrix);
     tmp_m = NEW_GF_MATRIX(n, k);
     /*
      * fill the matrix with powers of field elements, starting from 0.
